remove the integer number

+++ backend/models.py
@@ -13,7 +13,7 @@ class User(db.Model, SerializerMixin):
     id=db.Column(db.Integer, primary_key=True)
     name=db.Column(db.String(100), nullable=False)
     email=db.Column(db.String(100), unique=True)
-    contact=db.Column(db.Integer(50))
+    contact=db.Column(db.Integer)
     password=db.Column(db.String(50))
 
     enrollments=db.relationship("Enrollment", backref="user", lazy=True)
